requirements.txt
azure-functions
azure-ai-vision-imageanalysis>=1.0.0b1 
azure-core
openai>=1.0.0
azure-cognitiveservices-speech
python-dotenv 
PyJWT
cryptography
requests